requirements.txt
streamlit
mysql-connector-python
qrcode
Pillow
